Modifs mineures
Quelques modifs pour que ça compile sur ma version de gcc, j'pense que
ça marchera chez toi aussi du coup
J'ai viré les "restrict", faudra voir si on les remet pour le rendu

+++ ft_printf/srcs/ft_d.c
@@ -35,7 +35,7 @@ size_t	ft_va_arg_d_old(va_list ap, t_conv *list)
 	}
 	else
 	{
-		cpy = ft_memalloc((!(list->left) ? list->field : ft_strlen(cpy2)));
+		cpy = ft_memalloc((!(list->left) ? (size_t)list->field : ft_strlen(cpy2)));
 		while ((int)i < list->field - 1 && (!(list->left)))
 		{
 			cpy[i] = ' ';
